Move InputWithUnits and input syntax
